fix(test): centralize telemetry fixtures and fix scaffolding test regressions
